clean kb loop by resetting result variable
if get-kbupdate input is not only a single kb, you may receive the output from the first kb over and over again. to fix that the result variable in the foreach kb loop needs to be reset on every iteration, because it can be that you drag the result from the previous iteration within. The conditions inside the loop don't process the it well.