Add ability to download from link directly (#135)
Prior to this commit Save-KBUpdate processed and downloaded every link
associated with a particular requested KB from Get-KBUpdate. This led
to more files being downloaded than sometimes desired.

With this change it is now possible to supply a link directly to Save-KBUpdate
which will download only that file. You can pass multiple links to the -Link parameter
if you have multiple files you would wish to download from the results.